GO-3588 Add restrictions to archived objects
